Modified JNI function preprocessing to work around bug in CPS register allocator; added 'max jnigen' to separate JNI related preprocessing from image building

+++ VM/src/com/sun/max/vm/jdk/JDK_java_lang_Runtime.java
@@ -26,7 +26,7 @@ import com.sun.max.vm.heap.*;
 
 /**
  * Implements method substitutions for {@link java.lang.Runtime java.lang.Runtime}.
- * 
+ *
  * @author Bernd Mathiske
  */
 @METHOD_SUBSTITUTIONS(Runtime.class)
@@ -42,7 +42,7 @@ public final class JDK_java_lang_Runtime {
      */
     @SUBSTITUTE
     private long freeMemory() {
-        return Heap.reportFreeSpace().toLong();
+        return Heap.reportFreeSpace();
     }
 
     /**
